move demo application code to app_demo.c/.h
