+++ Software/Nucleo-F446RE-advanced-example/Core/Inc/app_demo.h
@@ -1,0 +1,8 @@
+#ifndef INC_APP_DEMO_H_
+#define INC_APP_DEMO_H_
+
+
+void app_demo_main(void);
+
+
+#endif /* INC_APP_DEMO_H_ */
